sub: make OSD size per-OSD-object
Because subtitles may be rendered at different resolution (consider
non-OSD subtitle rendering with vf_ass).

It may be a bit confusing that non-OSD subtitle rendering reuses the
OSD code path and that not all OSD objects have the same size as
consequence. But on the other hand, the OSD resolution can be considered
part of the renderer cache state. Additionally, at least the "old"
(OSDTYPE_SUBTITLE) subtitle renderer is not separate from the OSD code,
as well as the indexed-to-RGBA conversion.

Also, include display_par/video_par in resolution change detection too.