demux_lavf: do not prefix filename passed to libavformat with "mp:"
Opening files with the libavformat AVISynth demuxer ("avs"/avisynth.c)
fails, because the filename we pass to avformat_open_input() is prefixed
with "mp:". Normally, this doesn't matter, because data is read with the
stream interface. The AVISynth demuxer can't use this, because the
Avisynth API (apparently) can't read scripts from memory, and requires
a filename.

It's not clear why the existing code added the "mp:" prefix, and it was
added with the initial commit. Since this prefix doesn't seem to have
any purpose and prevents AVS playback from functioning, remove it.
Fixes #5.