demux_lavf: add support for libavdevice
libavdevice supports various "special" video and audio inputs, such
as screen-capture or libavfilter filter graphs.

libavdevice inputs are implemented as demuxers. They don't use the
custom stream callbacks (in AVFormatContext.pb). Instead, input
parameters are passed as filename. This means the mpv stream layer has
to be disabled. Do this by adding the pseudo stream handler avdevice://,
whose only purpose is passing the filename to demux_lavf, without
actually doing anything.

Change the logic how the filename is passed to libavformat. Remove
handling of the filename from demux_open_lavf() and move it to
lavf_check_file(). (This also fixes a possible bug when skipping the
"lavf://" prefix.)

libavdevice now can be invoked by specifying demuxer and args as in:

    mpv avdevice://demuxer:args

The args are passed as filename to libavformat. When using libavdevice
demuxers, their actual meaning is highly implementation specific. They
don't refer to actual filenames.

For some reason, "mpv avdevice://matroska:test.mkv" doesn't work,
although it should.

Note:

libavdevice is disabled by default. There is one problem: libavdevice
pulls in libavfilter, which in turn causes symbol clashes with mpv
internals. The problem is that libavfilter includes a mplayer filter
bridge, which is used to interface with a set of nearly unmodified
mplayer filters copied into libavfilter. This filter bridge uses the
same symbol names as mplayer/mpv's filter chain, which results in symbol
clashes at link-time.

This can be prevented by building ffmpeg with --disable-filter=mp, but
unfortunately this is not the default.

This means linking to libavdevice (which in turn forces linking with
libavfilter by default) must be disabled. We try doing this by compiling
a test file that defines one of the clashing symbols (vf_mpi_clear).

To enable libavdevice input, ffmpeg should be built with the options:

    --disable-filter=mp

and mpv with:

    --enable-libavdevice

Originally, I tried to auto-detect it. But the resulting complications
in configure don't seem worth the trouble. The original patch that
autodetects libavfilter/libavdevice in a way that disables libavdevice
use if a conflict would arise follows inline:

diff --git a/configure b/configure
index 1d6588f..d155bba 100755
--- a/configure
+++ b/configure
@@ -334,6 +334,7 @@ Optional features:
   --enable-rpath         enable runtime linker path for extra libs [disabled]
   --disable-libpostproc  disable postprocess filter (vf_pp) [autodetect]
   --disable-libavdevice  disable libavdevice demuxers [autodetect]
+  --disable-libavfilter  disable libavfilter [autodetect]

 Codecs:
   --enable-gif              enable GIF support [autodetect]
@@ -493,6 +494,7 @@ _ass=auto
 _libass_osd=auto
 _rpath=no
 libpostproc=auto
+libavfilter=auto
 libavdevice=auto
 _asmalign_pot=auto
 _stream_cache=yes
@@ -715,6 +717,8 @@ for ac_option do
   --disable-libpostproc) libpostproc=no ;;
   --enable-libavdevice) libavdevice=yes ;;
   --disable-libavdevice) libavdevice=no ;;
+  --enable-libavfilter) libavfilter=yes ;;
+  --disable-libavfilter) libavfilter=no ;;

   --enable-enca)        _enca=yes       ;;
   --disable-enca)       _enca=no        ;;
@@ -2855,11 +2859,60 @@ else
 fi

+# expect that pkg_config_add changes only: $extra_cflags $extra_ldflags
+lavfi_original_extra_cflags="$extra_cflags"
+lavfi_original_extra_ldflags="$extra_ldflags"
+
+echocheck "libavfilter"
+libavfilter_check_braindeath=no
+if test "$libavfilter" = auto ; then
+    libavfilter=no
+    if pkg_config_add "libavfilter >= 3.17.0" ; then
+        libavfilter=yes
+        libavfilter_check_braindeath=yes
+    fi
+fi
+
+if test "$libavfilter_check_braindeath" = yes ; then
+    avfilter_braindeath=no
+
+    cat > $TMPC <<EOF
+    #include <libavfilter/avfilter.h>
+    void vf_mpi_clear(void) {}
+    int main(void) { vf_mpi_clear(); avfilter_register_all(); return 0; }
+EOF
+
+    # The libavresample failure case is because libavfilter has some filters
+    # that use libavresample (while others use libswresample). But in these
+    # cases, the ffmpeg libavfilter .pc file fails to declare libavresample as
+    # dependency. If we don't explicitly link to libavresample, we will fail
+    # due to missing libavresample symbols.
+    if ! cc_check ; then
+        avfilter_braindeath=yes
+        libavfilter=no
+        res_comment="symbol clashes due to braindead libavfilter mplayer filter bridge (compile ffmpeg with --disable-filter=mp), or buggy ffmpeg .pc file missing libavresample"
+        extra_cflags="$lavfi_original_extra_cflags"
+        extra_ldflags="$lavfi_original_extra_ldflags"
+    fi
+fi
+if test "$libavfilter" = yes ; then
+    def_libavfilter='#define CONFIG_LIBAVFILTER 1'
+else
+    def_libavfilter='#undef CONFIG_LIBAVFILTER'
+fi
+
+echores "$libavfilter"
+
+
 echocheck "libavdevice"
 if test "$libavdevice" = auto ; then
     libavdevice=no
-    if pkg_config_add "libavdevice >= 54.0.0" ; then
-        libavdevice=yes
+    if test "$libavfilter" = no ; then
+        res_comment="libavfilter"
+    else
+        if pkg_config_add "libavdevice >= 54.0.0" ; then
+            libavdevice=yes
+        fi
     fi
 fi
 if test "$libavdevice" = yes ; then
@@ -3292,6 +3345,7 @@ LIBBS2B = $_libbs2b
 LCMS2 = $_lcms2
 LIBPOSTPROC = $libpostproc
 LIBAVDEVICE = $libavdevice
+LIBAVFILTER = $libavfilter
 LIBSMBCLIENT = $_smb
 LIBQUVI = $_libquvi
 LIBTHEORA = $_theora
@@ -3452,6 +3506,7 @@ $def_zlib

 $def_libpostproc
 $def_libavdevice
+$def_libavfilter

 /* Audio output drivers */