stream_lavf: handle https protocol
We don't have native support and probably never will.

This requires ffmpeg to be configured with --enable-openssl. It's not
autodetected.

The cache is not enabled for https. I'm not sure whether it's as easy as
enabling the internal cache, or if there are complications with using
ffmpeg streams with the cache. This will be investigated later.