stream: use ffmpeg for http streaming
The internal mplayer http implementation is terrible and old. Try to use
ffmpeg instead.

Enable the cache for ffmpeg streams. In theory, we want to enable the
cache for network streams only (or at least, not enable it for files),
but there doesn't seem to be any way to query this.

Should something turn up that shows that the mplayer implementation is
better than ffmpeg's, this commit might be reverted until ffmpeg is
ready. So far, the ffmpeg implementation seems to very work well though.

The original mplayer stream implementation is made available under
"mp_http://" and "mp_http_proxy://".

Note: "http_proxy://" doesn't work anymore. Use "httpproxy://" to get
the ffmpeg implementation of it.