core: automatically pause on low cache
When the cache runs below a certain threshold, automatically pause the
player. When the cache is filled again, unpause again.

This is intended to help with streaming from http. It's better to pause
a while, rather than exposing extremely crappy behavior when the stream
layer causes packet reads to block the entire player.

In theory, we should try to increase the cache if underruns happen too
often. Unfortunately, changing the cache implementation would be very
hard, because it's insane code (forks, uses shared memory and "volatile"
etc.). So for now, this just changes reduces the frequency of the
stuttering if the network is absolutely too slow to play the stream in
realtime.