input: don't generate double clicks for mouse wheel
Button 3 and 4 are mapped to the mouse wheel. Double click events for
them annoying and not useful at all.

I don't know about buttons 5-19; for all I know these could be mapped
to wheels as well. Even if not, double click events are probably not
very important for these. Disable double clicks for these as well.