osx: remove NSApplicationLoad hack
NSApplicationLoad is a hack for Carbon Applications to make the operating
system think they are Cocoa. It takes care of basic initialization (such as
connection to the window server). mpv was using this to get around the
entangled initialization process.

NSApplicationLoad also created a basic menu for the application which mpv then
customized from cocoa_common. This resulted in odd menu bugs, such as `cmd + q`
not working on some versions of OSX, double apple logos which was fixed with
yet another hack and more uglities.

The next commit will add the MainMenu back. mpv is now a nibless application.